Events, Fines and Notifications + Merge Conflicts solved

May god bless this merge...

+++ application/core/src/main/java/com/github/pieter_groenendijk/model/payment/PaymentStatus.java
@@ -13,7 +13,7 @@ public class PaymentStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "paymentStatusId")
-    private Long paymentStatusId;
+    private short paymentStatusId;
 
     @Column(
         name = "title",
